Complete

+++ app/controllers/orders_controller.rb
@@ -27,26 +27,34 @@ class OrdersController < ApplicationController
 
   def generates_order
     address = current_user.delivery_addresses.find_by(address: params[:address])
-    @order = current_user.orders.create!(
-      shipping_fee:         params[:shipping_fee].to_f,
-      total_payment:        params[:total_payment].to_f,
-      payment_method:       params[:payment_method],
-      delivery_address_id:  address.id,
-      cart_id:              current_user.cart.id,
-      status:               0
-    )
-    @cart_items.map do |cart_item|
-      @order.order_items.create!(
-        price_pay:   cart_item.product.price,
-        quantity:    cart_item.quantity,
-        total_price: cart_item.total_price,
-        product_id:  cart_item.product.id
+    if address
+      @order = current_user.orders.create(
+        shipping_fee:         params[:shipping_fee].to_f,
+        total_payment:        params[:total_payment].to_f,
+        payment_method:       params[:payment_method],
+        delivery_address_id:  address.id,
+        cart_id:              current_user.cart.id,
+        status:               0
       )
-    end
-    CartItem.where(cart_id: current_user.cart.id).delete_all
+      @cart_items.map do |cart_item|
+        @order.order_items.create!(
+          price_pay:   cart_item.product.price,
+          quantity:    cart_item.quantity,
+          total_price: cart_item.total_price,
+          product_id:  cart_item.product.id
+        )
+      end
+      CartItem.where(cart_id: current_user.cart.id).delete_all
 
-    redirect_to root_path
-    flash[:notice] = 'Order success!'
+      redirect_to root_path
+      flash[:notice] = 'Order success!'
+    else
+      @q                = Product.ransack(params[:q])
+      @order_items      = current_user.cart.cart_items.includes(:product).latest
+      flash[:alert] = "Address can't not empty"
+      render :new
+    end
+    
   end
 
   def generates_charge
